move quantity_as_xr to reporting.utils
